Incremental work on parallel messaging. Some of the receiving node code is now in place. This is the last checkin with the dummy messaging, next round will use MPI. -- Upi
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@23 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 7f7420dbf6b443f1637df2df07bc91e076d94368

+++ scheduling/ClockTickWrapper.cpp
@@ -18,9 +18,9 @@ Finfo* ClockTickWrapper::fieldArray_[] =
 ///////////////////////////////////////////////////////
 // Field definitions
 ///////////////////////////////////////////////////////
-	new ValueFinfo< int >(
+	new ValueFinfo< double >(
 		"stage", &ClockTickWrapper::getStage, 
-		&ClockTickWrapper::setStage, "int" ),
+		&ClockTickWrapper::setStage, "double" ),
 	new ValueFinfo< double >(
 		"nextt", &ClockTickWrapper::getNextt, 
 		&ClockTickWrapper::setNextt, "double" ),
